+++ src/main/java/zd/zero/waifu/motivator/plugin/personality/core/IdlePersonalityCore.kt
@@ -1,4 +1,4 @@
-package zd.zero.waifu.motivator.plugin.personality
+package zd.zero.waifu.motivator.plugin.personality.core
 
 import zd.zero.waifu.motivator.plugin.ProjectConstants
 import zd.zero.waifu.motivator.plugin.assets.VisualMotivationAssetProvider
@@ -6,6 +6,7 @@ import zd.zero.waifu.motivator.plugin.assets.WaifuAssetCategory
 import zd.zero.waifu.motivator.plugin.motivation.VisualMotivationFactory
 import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEvent
 import zd.zero.waifu.motivator.plugin.onboarding.UpdateNotification
+import zd.zero.waifu.motivator.plugin.personality.core.emotions.Mood
 import zd.zero.waifu.motivator.plugin.tools.doOrElse
 
 class IdlePersonalityCore : PersonalityCore {
