+++ src/main/java/zd/zero/waifu/motivator/plugin/personality/core/PersonalityCore.kt
@@ -1,6 +1,7 @@
-package zd.zero.waifu.motivator.plugin.personality
+package zd.zero.waifu.motivator.plugin.personality.core
 
 import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEvent
+import zd.zero.waifu.motivator.plugin.personality.core.emotions.Mood
 
 interface PersonalityCore {
     fun processMotivationEvent(
