+++ src/main/java/zd/zero/waifu/motivator/plugin/personality/core/emotions/NegativeEmotionDerivationUnit.kt
@@ -1,7 +1,8 @@
-package zd.zero.waifu.motivator.plugin.personality
+package zd.zero.waifu.motivator.plugin.personality.core.emotions
 
 import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEvent
 import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEvents
+import zd.zero.waifu.motivator.plugin.tools.ProbabilityTools
 import zd.zero.waifu.motivator.plugin.settings.WaifuMotivatorState
 import zd.zero.waifu.motivator.plugin.tools.toStream
 import java.util.stream.Stream
