+++ src/main/java/zd/zero/waifu/motivator/plugin/personality/core/emotions/PositiveEmotionDerivationUnit.kt
@@ -1,8 +1,9 @@
-package zd.zero.waifu.motivator.plugin.personality
+package zd.zero.waifu.motivator.plugin.personality.core.emotions
 
 import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEvent
 import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEventCategory
 import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEvents
+import zd.zero.waifu.motivator.plugin.tools.ProbabilityTools
 import zd.zero.waifu.motivator.plugin.settings.WaifuMotivatorState
 import java.util.stream.Stream
 import kotlin.random.Random
