+++ src/test/java/zd/zero/waifu/motivator/plugin/personality/NegativeEmotionCoreTests.kt
@@ -9,7 +9,9 @@ import zd.zero.waifu.motivator.plugin.alert.AlertConfiguration
 import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEvent
 import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEventCategory
 import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEvents
-import zd.zero.waifu.motivator.plugin.personality.NegativeEmotionDerivationUnit.Companion.OTHER_NEGATIVE_EMOTIONS
+import zd.zero.waifu.motivator.plugin.personality.core.emotions.EmotionCore
+import zd.zero.waifu.motivator.plugin.personality.core.emotions.Mood
+import zd.zero.waifu.motivator.plugin.personality.core.emotions.NegativeEmotionDerivationUnit.Companion.OTHER_NEGATIVE_EMOTIONS
 import zd.zero.waifu.motivator.plugin.settings.WaifuMotivatorState
 import zd.zero.waifu.motivator.plugin.tools.toList
 import kotlin.random.Random
